Move logic into dedicated files
